django/classnotes/019_Pagination_Filters_Search/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-decouple==2.1
django-filter==22.1
djangorestframework==3.13.1
Faker==14.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
